fix(get): do not allow `clean` to delete `get`
"Get" is how we fundamentally interact with all our code, so it doesn't
make sense to remove it, since cleaning is usually done to facilitate
getting other code files.

+++ get.lua
@@ -224,7 +224,9 @@ elseif scriptName == "display_disk" then
 elseif scriptName == "clean" then
     local fileList = fs.list("/disk")
     for _, file in ipairs(fileList) do
-        fs.delete("/disk/" .. file)
+        if file ~= "get" then
+            fs.delete("/disk/" .. file)
+        end
     end
     term.setTextColor(colors.orange)
     print()
